Address audit iteration 8 findings on cached-token contract
ClaudeSessionWorker.EmitMetrics used the pre-contract formula
`fresh = InputTokens - CachedInputTokens`, which under the new
fresh+cache_creation snapshot semantics clamped to 0 in the common
warm-cache case. Compute fresh directly from snapshot.InputTokens and
record the operator-facing total as InputTokens + CachedInputTokens.
Update ClaudeSessionTurnMetrics docstring to match.

Switch WorkItemUsageAggregator accumulators (and the WorkItem*Usage
record fields) to long so multi-row aggregates can sum past int.MaxValue
without silent overflow.

Add a one-shot SQLite migration (usage_contract_version column) that
subtracts cached_input_tokens from input_tokens on pre-fix Claude and
OpenCode rows so the aggregator's new InputTokens+CachedInputTokens sum
does not double-count legacy data; codex rows are no-ops because their
cached column was always 0.

Normalise GeminiStreamParser so AgentStreamSummary.InputTokens is the
fresh-only billable bucket across providers (Google's promptTokenCount
includes cachedContentTokenCount, matching codex's fresh-only contract).

Update the AcpClaudeTransport round-trip test that still pinned the old
snapshot contract.

CodeyBox-Prompt-Revision: 1

+++ src/CodeyBox.Core/WorkItemUsage.cs
@@ -3,25 +3,29 @@ namespace CodeyBox.Core;
 /// <summary>
 /// Token usage and estimated cost for a single iteration of a work item.
 /// Returned on the API and webhook surfaces as the <c>usage</c> block.
+/// Token counts are <see langword="long"/> so multi-row aggregates can exceed
+/// <see cref="int.MaxValue"/> without silent overflow (a single agent CLI call
+/// already returns up to ~10⁷ tokens, and busy work items accumulate many calls).
 /// </summary>
 public sealed record WorkItemIterationUsage(
     int Iteration,
-    int TokensInput,
-    int TokensOutput,
-    int TokensReasoning,
-    int TokensCached,
+    long TokensInput,
+    long TokensOutput,
+    long TokensReasoning,
+    long TokensCached,
     double CostUsd,
     long ElapsedMs);
 
 /// <summary>
 /// Cumulative token usage and estimated cost across every iteration of a work item.
 /// Returned on the API and webhook surfaces as the <c>usageTotal</c> block.
+/// Token counts are <see langword="long"/>; see <see cref="WorkItemIterationUsage"/>.
 /// </summary>
 public sealed record WorkItemUsageTotal(
-    int TokensInput,
-    int TokensOutput,
-    int TokensReasoning,
-    int TokensCached,
+    long TokensInput,
+    long TokensOutput,
+    long TokensReasoning,
+    long TokensCached,
     double CostUsd,
     long ElapsedMs);
